work in progress
Former-commit-id: 7d19730ec08336cf623d9763cb940f9a1a4a8c8b
Former-commit-id: 30bb609a8d713960d7125453278e8e83c497a784

+++ src/tex.c
@@ -635,7 +635,7 @@ static texp tex_black_and_white (SDL_Surface *in,
 
             c = getPixel(in, ix, iy);
             
-            uint8_t avg = (c.r + c.g + c.b) / 3;
+            uint8_t avg = ((int)c.r + (int)c.g + (int)c.b) / 3;
             c.r = avg;
             c.g = avg;
             c.b = avg;
